# client/src/pages/Swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CloudXi Museum",
    "description": "This is the REST API documentaion for the Cloudxi Museum online management system",
    "version": "1.0.0",
    "contact": {
      "email": "CloudXiMuseum@gmail.com"
    }
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
...
  "paths": {
...
    "/employee/check-in-out/{ticketCode}": {
      "get": {
        "tags": [
          "Check in & Check out"
        ],
        "summary": "Get the status of a ticket",
        "description": "Get the status of a ticket",
        "parameters": [
          {
            "name": "ticketCode",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/employee/check-in-out/check-in": {
      "post": {
        "tags": [
          "Check in & Check out"
        ],
        "summary": "Checks a ticket in",
        "description": "Checks a ticket in",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/employee/check-in-out/check-out": {
      "post": {
        "tags": [
          "Check in & Check out"
        ],
        "summary": "Checks a ticket out",
        "description": "Checks a ticket out",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/employee/modify/{tableName}/insert": {
      "post": {
        "tags": [
          "Modify"
        ],
        "summary": "Inserts a new tuple to a table",
        "description": "Inserts a new tuple to a table",
        "parameters": [
          {
            "name": "tableName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/employee/modify/{tableName}/delete": {
      "post": {
        "tags": [
          "Modify"
        ],
        "summary": "Delete a tuple from a table",
        "description": "Delete a tuple from a table",
        "parameters": [
          {
            "name": "tableName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/employee/modify/{tableName}/modify": {
      "post": {
        "tags": [
          "Modify"
        ],
        "summary": "Modify a tuple in a table",
        "description": "Modify a tuple in a table",
        "parameters": [
          {
            "name": "tableName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/admin/modify/{tableName}/insert": {
      "post": {
        "tags": [
          "Modify"
        ],
        "summary": "Inserts a new tuple to a table",
        "description": "Inserts a new tuple to a table",
        "parameters": [
          {
            "name": "tableName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/admin/modify/{tableName}/delete": {
      "post": {
        "tags": [
          "Modify"
        ],
        "summary": "Delete a tuple from a table",
        "description": "Delete a tuple from a table",
        "parameters": [
          {
            "name": "tableName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/admin/modify/{tableName}/modify": {
      "post": {
        "tags": [
          "Modify"
        ],
        "summary": "Modify a tuple in a table",
        "description": "Modify a tuple in a table",
        "parameters": [
          {
            "name": "tableName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  }
}